+++ src/transactions/transactionBuilder/index.ts
@@ -1,6 +1,7 @@
 // Copyright © Aptos Foundation
 // SPDX-License-Identifier: Apache-2.0
 
+export * from "./abiJson";
 export * from "./helpers";
 export * from "./transactionBuilder";
 export * from "./remoteAbi";